refactor: combine resouceMapOwner into resourceOwner middleware
