+++ src/routes/ResourceMapRoutes.ts
@@ -3,7 +3,7 @@ import express, {Request, Response, NextFunction} from 'express';
 import ResourceMapController from '../controllers/ResourceMapController';
 import Container from '../utils/container';
 import AuthenticationMiddleware from '../middlewares/authentication';
-import ResourceMapOwnerMiddleware from '../middlewares/resourceMapOwner';
+import ResourceOwnerMiddleware from '../middlewares/resourceOwner';
 import roleValidator, {ADMINS, TEACHERS} from '../middlewares/authorization';
 
 export default () => {
@@ -12,13 +12,15 @@ export default () => {
     Container.getInstance().get('ResourceMapController');
   const authenticationMiddleware: AuthenticationMiddleware =
     Container.getInstance().get('AuthenticationMiddleware');
-  const resourceMapOwnerMiddleware: ResourceMapOwnerMiddleware =
-    Container.getInstance().get('ResourceMapOwnerMiddleware');
+  const resourceOwnerMiddleware: ResourceOwnerMiddleware =
+    Container.getInstance().get('ResourceOwnerMiddleware');
 
   const auth = (req: Request, res: Response, next: NextFunction) =>
     authenticationMiddleware.authentication(req, res, next);
-  const resourceMapsOwner = resourceMapOwnerMiddleware.resourceMapsOwner;
-  const resourceMapOwner = resourceMapOwnerMiddleware.resourceMapOwner;
+  const resourceMapsOwner = (req: Request, res: Response, next: NextFunction) =>
+    resourceOwnerMiddleware.resourceMapsOwner(req, res, next);
+  const resourceMapOwner = (req: Request, res: Response, next: NextFunction) =>
+    resourceOwnerMiddleware.resourceMapOwner(req, res, next);
 
   resourceMapRouter.post(
     '/bulkCreate',
